requirements.txt
asgiref==3.8.1
Django==5.0.4
django-filter==24.2
djangorestframework==3.15.1
Markdown==3.6
sqlparse==0.4.4
tzdata==2024.1
